backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
httpx>=0.27.0
boto3>=1.34.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
mangum>=0.17.0

langchain==0.2.16
langchain-aws>=0.1.0
langchain-community>=0.2.0
